requirements.txt
# Web Framework & Server
Django==5.0.4
gunicorn==21.2.0
whitenoise[brotli]==6.6.0
setuptools==69.5.1
wheel==0.43.0

# Database & Environment
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.5
requests==2.31.0

# AI & LLM (Switched to latest google-genai as requested)
google-genai==0.4.0
pinecone==3.2.2

# Data Analysis & Processing
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
asteval==0.9.31
openpyxl==3.1.2

# File Parsing
pypdf==4.1.0
pdfplumber==0.11.0
python-docx==1.1.0

# Project Utilities
django-ratelimit==4.1.0
django-cryptography==1.1
